requirements.txt
python-dotenv
openai
gspread
oauth2client
flask